kernel: simplify notify() adding common popnote() function
Handlin notes is common for all architectures
except how the note has to be pushed on the user
stack.

This change adds a popnote() function that returns
only the note string or nil if the process should
not be notified (no notes or user notes hold off).

Popnote() also handles common errors like notify
during note handling or missing note handler and
will suicide the process in that case.