devip: do tcp mss clamping when forwarding packets
when forwarding packets (gating), unconditionally
check tcp-syn packets for the mss-size option and
reduce it to fit the mtu of the outgoing interface.

this is done by exporting a new tcpmssclamp() function
from ip/tcp.c  that takes an ip packet and its buffer size
and the effective mtu of the interface and adjusts
the mss value of tcp syn options.

this function is now also used by devbridge, enforcing
a tcp mss below the tunnel mtu.