arp: make sure arpresolve() only returns single packet
Due to locking changes, it is possible for arpresolve()
to return multiple packets in the multicast case,
resulting in etherbwrite() to fail assert(bp->list == nil).

Ensure that arpresolve() always returns the first packet
from the hold chain and frees the rest if any.

For arpenter(), we want to transmit the whole chain of
packets, so we detach them from the arp entry before
calling arpresolve() and ignoring its result.