nusb/audio: add support for USB audio 2.0
USB audio 2.0 mandates the use of an interface association descriptor
to group together the interfaces that comprise an audio function, and
the AudioControl interface no longer lists the streams. So for 2.0
devices, find and use the IAD to enumerate the streams.

Additionally, we now need to locate the terminal associated with the
stream, and the clock for that terminal. This replaces the frequency
fields in the type I format type descriptor, and the sampling rate is
now set with a request directed at the clock entity in the control
interface, rather than the endpoint.