cwfs: remove noauth and nonone commans from fileserver console
The noauth and nonone commands are only valid in config mode.
The problem is that some people assumed that issuing the
noauth command in the runtime fileserver console would be
persistent across reboots, which is not the case.

To avoid this confusion, we remove these commands from the
fileserver console.

To still give a way to disable authentication at runtime,
define a authdisabled flag that can be toggled at runtime.