sdmmc: new interface for SDio
We want to support the internal emmc device on the reform,
which has a different initialization sequence from SDcard.

The problem is, all mmc controller drivers where using
the command index to derive the command properties.
The command index's interpretation depends on the
command set. And MMC has different commands then SD.
Also, there are the APP_CMD escaped commands which
drivers then tried to recover the state...

All of this is total nonsense... The controller drivers
should not care and the command properties should
be maintained by port/sdmmc.c.

So we pass the command as a struct SDiocmd, which
has all the properties (command index, response type,
data transfer mode... and even a string for debugging).

The controller just converts this into register values
and just executes the commands.

Next, the controller drivers shouldnt snoop on the
commands and then try to apply bus and frequency
switching on their own. This is now made explicit by
having SDio.bus(io, width, speed) function.