sdmmc: implement boot partition access for emmc
We expose the boot partitions as separate units,
such as sdM0 is the user data partition and sdM1
is the first boot partition and sdM2 is the
second boot partition.

On access, we automatically switch the
BOOT_PARTITOIN_ACCESS bits in the EXT_CSD.

Also, the units ctl file contains "boot enabled"
or "boot disabled" if that unit is the active
or inactive boot partition.