libmemdraw: don't return Buffer copy from calc functions
The calc functions get their buffers passed by value.
This is convenient as the code usually modifies the
buffers during iteration.

However, making ANOTHER copy (odst) and returning it
at the end is a bit silly. We already made a copy
when passing the arguments, and the caller of the
calc function can just reuse the copy it already has.

So changing the return type from Buffer to void.