dossrv: Implement support for != 512 sector and track sizes
The iotrack buffer layer always assumed 512 byte sector
size and 9 sectors per track.

To support 4K sector size fats, make the iotrack code
deal with it.

Instead of the fixed size Track structure, we just allocate
the pointers and buffers dynamically, and move the sector
size and tack size in the Xfs structure.

Tracks can still be reused between differnet file-systems
after a purgetrack().

The initial fat header has to be read with devread()
instead of getsect() to determine the sector/track size
of the file-system.