dhcp: fix out of bounds access in "ANDROID_METERED" fix
the previous change introduces a out of bounds access
as it does not change n.

it is also conceptually wrong because this routine is
supposed to just verify the structure. as later getopts()
is *NOT* going to deal with malfored TLV's.

this actually replaces the android magic garbage with
OBpad bytes, which getopts() later will handle correctly
and makes sure the garbage is fully contained within
the buffer boundaries.

thanks sigrid for testing.