ip/ipconfig: initial dhcpv6 support, clean default-routes and /net/ndb on remove
This adds a very basic (probably wrong) DHCPv6 client,
to handle the "managed"-flag in IPv6 router solicitations.

We add -U option to pass the DHCPv6 client id as well as
an -s flag to manually add a dns server (because ppp
is going to call ipconfig to handle all the configuration
and write-back to /net/ndb in the future).

Have the remove command also remove default routes and
/net/ndb entries. (needed by ppp).