ip/ppp: just use mtu requested from the peer.
the mtu negotiation seems wrong. it seems to assume
that the peer requests a mtu to us, but actually this
doesnt seem to be the case and if the peer doesnt
request it, we'd always use Defmntu of 1450, no
matter what we set on with -m option.

instead, we use the mtu specified with -m option
and always request it to the peer. if the peer requests
a different mtu we use that instead.

this at least gives some control and allows to just use
the correct mtu from pppoe (1492).