ndb/query: add -i and -c flags unifying the functionality of ndb/ipqeury
I'm a bit tired of ndb/ipquery having a different interface
than ndb/query. Given that most scripts are running it in a sed
pipeline to extact just the value without the attribute name
while ndb/query does the thing one would expect.

So this allows one to do ipinfo searches (also thru ndb/cs
using the -c flag) while only getting the values (when only
one rattr is specified). And honiuring the the -a and -m
flags as well.