devip: generate more reasonable laddr for UDP "headers" connection writes
By default, we where just selecting a source
address based on the remote unicast address.

But for bcast/mcast, it makes more sense to
use the interface address (which was previously
generated by ipv6local())
to ensure that the reply is sent to the same
interface that the request came in from,
and make link-local remotes work.

this is analog to the address selection used
for "connection" based udp listener.

This code only triggers when the user has
not changed the source address to a unicast
address in the multicast case.