9boot: implement *nocga= and *nokbd= kernel boot parameters
Some embedded pcs like the pcengine apu2 has a bios
that emulates cga and ps2 keyboard using the serial
console.

This leads to doubling of characters in the bootloader
console, as the code would output on both, the uart
and on the cga "screen".

To avoid this, *nocga= boot parameter is added,
which essentially disables the cga text mode. This
boot parameter might also be usefull for no-legacy
uefi machines that generally have no cga text mode
anymore.

Second, we add *nokbd= which makes both the kernel and
9boot stop accessing the ps2 keyboard.