7c: more registers, improve constant propagation, implement load pipelining
use all 15 (before was 7) registers (R9-R23)
for variable registerization.

don't replace $0 integer constant with register
in constprop(), as we have a zero-register (R31) on arm64,
which the linker is able to encode just fine.

handle floating-point constants in constprop().

when we have a load immediately followed by
the load-dependent instruction, see if we
can find another in-dependent instruction to
put after the load. this results in nice
pipelined code and also clusters loads and
stores (which we could later optimize with
MOVP instructions). this optimization gave
me around 20% speedup for sha2 on cortex a53.