bcm, bcm64, imx8, kw: Fix missing preempted() call on non-clock interrupt (thanks Anthony Martin)
As Anthony Martin noticed, some of the newer ARM and ARM64
kernels missed the preempted() call after handling non-clock
interrupts, which resulteds in a delay of scheduling higher
priority proc that has been woken up by the interrupt handler
(such as a kproc for attending some device or a higher
priority process that waited on some device to complete).

These kernels are all multiprocessors, so the woken process
would most likely be run by any other cpu *eventually*...
... with some delay.

Anyway, this change fixes it.

The change on zynq is cosmetic only. It was not affected
by this problem.