nusb/kb: fix portreset feature and recovery
portreset feature cannot work with xhci, as we cannot
keep the address.

if nusb/kb encounters a transaction error, request a
port reset and exit immediately.

then usbd will detach the device from the port and
and try a re-attach of the device when still present,
which automatically triggers a port reset.