ether82563: clear force speed/duplex bits in link-up (thanks taw)
we should not bring the link up in detach(), this makes no sense.

bring the link up in attach(), but make sure prior link-speed
force bits are clear, so that the information is properly
taken from the phy as it negotiates the link speed.

tested with i82574 and i82579, and fixes link not working
with anything but 1000baseT. thanks taw for your patence!