upas/fs: write index out to temp file
pridx() can be extremely slow when reading from
imap, because it ensures each message is cached.
Ensuring each message is cached may take a round
trip to the IMAP server.

Instead, we should write the index to a temp file,
and rename it, to reduce the window wher the index
can be broken.