7c: fix constant propagation for negative 64-bit constants
the issue is that MOVW $imm, R does not sign extend to 64-bit,
leaving the upper 32 bit of the register as zeros.

this means we have to be carefull when propagating negative
constants across loads.

add a "as" argument for the move instruction type,
we can always substitute if we have the same move operation,
then, we test for special cases where it doesnt matter.

to reproduce the issue:

void
main(void)
{
	long foobar = -1;
	vlong x = -1;
	print("%d\n", x == -1);
}

with the change:

/fd/5:2,9 - /fd/4:2,10
  	RETURN
  	TEXT	main+0(SB),0,$32
  	MOVW	$-1,R1
+ 	MOV	$-1,R3
  	MOV	$.string<>+0(SB),R0
- 	CMN	$1,R1
+ 	CMN	$1,R3
  	BNE	3(PC)
  	MOVW	$1,R5
  	B	2(PC)