ndb/mkhash: handle database= reordering, check offset overflow, remove file
when opening /lib/ndb/local, the db returned is the first entry
in the database= tuple, not the file passed to ndbopen(),
so search for the file in question.

make sure the file-offset doesnt exceed the NDBSPEC bit (23-bit).

remove the created hash file on error, and set TMP bit.