git/fs: don't unload large objects when inserting into cache
When inserting an object into cache, we try to remove objects
from cache to keep the size below the limit. If the object is
larger than the cache, we try to unload not only every cached
object, but the object we're in the process of inserting.

This leads to an immediate unload of the object, which is very
wrong.