kernel: get rid of mountid and quadratic algorithms in pgrpcpy() and devproc/readns1()
Ori discovered that the 64-bit Mount.mountid was
handled incorrectly in devproc's readns1() as Chan.nrock
is 32-bit integer.

Instead of sorting the Mount entries by 64-bit mountid
using pgrpinsert(), always maintain orderd linked list in
Pgrp.mntorder.

For pgrpcpy(), we also add a Mount *norder "forward"
pointer, which we temporarily stuff new Mount's into
the original pgrp Mount's. A the end, we re-insert
all the new mounts using pgrpinsert().

Pgrpinsert() is now O(1) as it just needs to append
to the list (and we keep a tail link pointer).

Add Mhead *umh into Mount struct and devproc's readns1()
becomes trivial.

The 64-bit mountid is not needed anymore, so remove
as well as the nextmount() generator.