kernel: refactor chdev(1) machinery
As the "notallowed" bitmap is part of Pgrp,
the code dealing with it should reside in
pgrp.c as well.

Rename Pgrp.notallowed[] to Pgrp.devmask[]
and make it an array of uchar[]. This
avoids the 8 * sizeof(pgrp->devmask[0])
for generating the shift counts.

Change devallowed() to devmasked() to be
symmetric with devmask().

Put functions in lexical order in portfns.h.

Make devmasked() take a devtab index instead
of device character, avoiding calling devno()
twice and makes it applicable to be used
in a loop (such as when reading #c/drivers).

Simplify devmask(): Convert to bitmap
and afterwards invert it while we copy into
Pgrp->devmask[] under lock using xor.

Move the inheritance logic into pgrpcpy()
and do it under the ns lock. This way,
a background forker will not inherit
a partially changed devmask. This only
impatcs rforking with CNAMEG flag.