libc: mark exits as profile-able
Exits had been marked as noprof, this was likely done because
there would be some noise added by _profdump itself (added as an onexit handler).
However, knowing if your atexit handlers are taking up a lot of time is useful,
and its possible to all but remove the impact by stopping the clock before we call things
like create and print.