ircrc: revert 78e2f69c0d7e4c17227e1ec0545cbb2a0ae1e96e and introduce a -u flag for username
ZNC completely ignores NICK when choosing the network to connect to,
so we don't need to allow / in nicknames (besides that rfc 1459 and
2812 disallow it).  This diff instead changes -u to take a username
and makes -U the flag for disabling encryption.  The user can then
specify their bouncer account password in PASS, and pass their bouncer
account username/network in USER.  This is compatible with ZNC, soju,
and pounce and works with factotum too as the same keyspec can be used
for all the networks that the user configured on their bouncer as long
as they pass the same nickname to ircrc for each network connection.