devip: handle announce and connect ctl error (thanks panini)
When announce and connect handlers fail,
set the state of the connection back to Idle,
allowing retry without having to allocate
a new connection.

Before, the connection would be stuck in
"Announcing"/"Announced" or "Connecting"/"Connected"
states and refuse to accept another connect
or announce ctl message.

This makes it easier to search for a free port
by just retrying announce ctl message.